requirements.txt
requests>=2.7.0
pymongo>=3.2.2
matplotlib>=1.4.3
Pillow>=3.2.0
